fix(psalm phase1): clean up invalid returns, union return type for Twig filters, and redundant casts

+++ src/Service/Integration/Jira/JiraOAuthApiService.php
@@ -350,7 +350,7 @@ class JiraOAuthApiService
             $this->delete(sprintf(
                 'issue/%s/worklog/%d',
                 $sTicket,
-                (int) ($entry->getWorklogId() ?? 0)
+                $entry->getWorklogId() ?? 0
             ));
 
             $entry->setWorklogId(null);
